Move files back to primary folder
